requirements.txt
SpeechRecognition
PyAudio
setuptools
google-cloud-aiplatform
google-cloud-texttospeech